+++ src/composables/useOpenMeteo.ts
@@ -9,48 +9,59 @@ export function useOpenMeteo() {
     const BASE_URL = `https://api.open-meteo.com/v1/forecast`;
 
     const getWeatherData = async (
-        location: Location
+        location: Location,
+        temprature?: string,
+        showHistoricalData?: string,
+        forecastDays?: number // by default 7 as in open-meteo
     ) => {
         weatherData.value = []
 
         const startDate = new Date();
-            startDate.setDate(startDate.getDate() - 7);
-            
-            const endDate = new Date();
-            endDate.setDate(endDate.getDate() + 7);
+        if (showHistoricalData) {
+            startDate.setDate(startDate.getDate() - (forecastDays || 7));
+        }
 
-            const params = new URLSearchParams({
-                latitude: location.latitude,
-                longitude: location.longitude,
-                hourly: "temperature_2m",
-                start_date: startDate.toLocaleDateString('en-CA'),
-                end_date: endDate.toLocaleDateString('en-CA'),
-                temperature_unit: "celsius"
-            });
+        const endDate = new Date();
+        endDate.setDate(endDate.getDate() + (forecastDays || 7));
 
-            const response = await fetch(`${BASE_URL}?${params.toString()}`);
-            
-            if (!response.ok) {
-                return
-            }
+        const params = new URLSearchParams({
+            latitude: location.latitude,
+            longitude: location.longitude,
+            hourly: "temperature_2m,precipitation",  //precipitation can be: rain, showers, snow
+            start_date: startDate.toLocaleDateString('en-CA'),
+            end_date: endDate.toLocaleDateString('en-CA'),
+            temperature_unit: temprature ?? "celsius",
+        });
 
-            const responseInJson = await response.json();
-            weatherData.value = mapDateToTempratue(responseInJson)
-            
+        const response = await fetch(`${BASE_URL}?${params.toString()}`);
+
+        if (!response.ok) {
+            return
+        }
+
+        const responseInJson = await response.json();
+        weatherData.value = mapDateToTempratue(responseInJson)
+
     }
 
     const mapDateToTempratue = (data: any): Weather[] => {
         const now = new Date();
         const weatherData: Weather[] = [];
-        
+
         for (let i = 0; i < data.hourly.time.length; i++) {
-            const date = new Date(data.hourly.time[i]);
+            const currentDate = new Date(data.hourly.time[i]);
+            const date = currentDate.toLocaleString('en-US', {
+                weekday: 'short',
+                hour: 'numeric',
+                hour12: true
+            });
             const temperature = new Temperature(data.hourly.temperature_2m[i]);
-            const isForecast = date > now;
-            
-            weatherData.push(new Weather(date, temperature, isForecast));
+            const rainChances = data.hourly.precipitation[i] || 0
+            const isForecast = currentDate > now;
+
+            weatherData.push(new Weather(date, temperature, isForecast, rainChances));
         }
-        
+
         return weatherData;
     }