+++ src/domain/entities/Weather.ts
@@ -4,6 +4,7 @@ export class Weather {
     constructor(
         public readonly date: Date,
         public readonly temperature: Temperature,
-        public readonly isForecast: boolean
+        public readonly isForecast: boolean,
+        public readonly rainChances: number
     ) {}
 }