+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 3. Add location entity for latitude and longitude
 4. Display weather data as grid
     1. Add tailwind Css
+    2. Add table layout
+    3. Show More in case more then 1 day of data